# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Extrittio API",
    "description": "IoT Hub platform REST API",
    "contact": {
      "name": "Extrittio contributors"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Authenticate and obtain a JWT token.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get the currently authenticated user.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "issue_token": {
            "type": "boolean",
            "description": "Return a bearer token for non-browser clients. Browser clients should\nuse the HttpOnly session cookie and leave this false."
          },
          "password": {
            "type": "string"
          },
          "tenant_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "user"
        ],
        "properties": {
          "token": {
            "type": [
              "string",
              "null"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
...
      "RoleSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "is_system"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "is_system": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "roles",
          "permissions",
          "permission_version"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "permission_version": {
            "type": "integer",
            "format": "int32"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "role": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            }
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT bearer authentication for explicitly enabled non-browser clients"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "extrittio_session",
        "description": "HTTP-only browser session cookie issued by POST /api/v1/auth/login"
      }
    }
  },
...
}